backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
asyncio==3.4.3
attrs==24.3.0
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
Django==5.1.5
django-cors-headers==4.6.0
djangorestframework==3.15.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
yarl==1.18.3
